# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  description: |
    REST API of the Bug Report System, a minimalistic Jira-like bug tracker.

    **Authentication.** The API uses a session cookie and CSRF protection:
    1. Call `GET /api/csrf` to get a CSRF token (this also sets the `JSESSIONID` cookie).
    2. Sign in with `POST /api/auth/login` (form fields `username` and `password`, where `username` is the email),
       sending the token in the `X-CSRF-TOKEN` header.
    3. Send the session cookie with every request, and the `X-CSRF-TOKEN` header with every `POST`, `PATCH` and `DELETE`.
       The token changes when you sign in, so call `GET /api/csrf` again after login and use the new token.

    Only `GET /api/csrf` and `POST /api/accounts` (registration) are public. Missing sign-in answers 401,
    a missing role, ownership rule or CSRF token answers 403. Errors have the body `{"message": "..."}`.
  title: Bug Report System API
  version: 0.1.0-SNAPSHOT
servers:
- description: Generated server url
  url: http://localhost:8089
...
paths:
...
  /api/comments:
    get:
      description: "Newest first; empty if there are none. `reportId` is required.\
        \ Access: any signed-in user."
      operationId: getComments
      parameters:
      - description: Id of the bug report.
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        in: query
        name: reportId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentResponse"
          description: "Comments of the report, newest first."
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "Validation failed, the JSON is malformed, or a path or query\
            \ value (for example an id) is malformed."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: Not signed in.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: A referenced resource does not exist.
      summary: List the comments of a report
      tags:
      - Comments
    post:
      description: "Adds a comment as the signed-in user. Returns 409 if the report\
        \ is closed. Access: any signed-in user."
      operationId: create_1
      parameters:
      - description: "CSRF token from `GET /api/csrf`. The token changes when you\
          \ sign in, so fetch a new one after login."
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCommentResponse"
          description: Comment created.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "Validation failed, the JSON is malformed, or a path or query\
            \ value (for example an id) is malformed."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: Not signed in.
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "The role or ownership check failed, or the CSRF token is missing."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: A referenced resource does not exist.
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: The request conflicts with a business rule or with existing
            data.
      summary: Add a comment
      tags:
      - Comments
  /api/comments/{commentId}:
    delete:
      description: "Access: signed-in user; the service further requires the ADMIN\
        \ role or being the comment's author. Returns 409 if the report is closed."
      operationId: delete
      parameters:
      - description: Id of the comment.
        example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        in: path
        name: commentId
        required: true
        schema:
          type: string
          format: uuid
      - description: "CSRF token from `GET /api/csrf`. The token changes when you\
          \ sign in, so fetch a new one after login."
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Comment deleted.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "Validation failed, the JSON is malformed, or a path or query\
            \ value (for example an id) is malformed."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: Not signed in.
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "The role or ownership check failed, or the CSRF token is missing."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: A referenced resource does not exist.
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: The request conflicts with a business rule or with existing
            data.
      summary: Delete a comment
      tags:
      - Comments
...
components:
  schemas:
    ApiErrorResponse:
      type: object
      description: JSON body of every error response.
      properties:
        message:
          type: string
          description: Human-readable description of the error.
          example: Request contains invalid values.
      required:
      - message
...
    CommentResponse:
      type: object
      description: "A comment in `GET /api/comments?reportId=...`."
      properties:
        authorId:
          type: string
          format: uuid
          description: Id of the author.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        authorName:
          type: string
          description: Display name of the author.
          example: Rachel Reporter
        bugReportId:
          type: string
          format: uuid
          description: Id of the commented report.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        content:
          type: string
          description: Comment text.
          example: This also happens in a private browser window.
        createdAt:
          type: string
          format: date-time
          description: When the comment was written (ISO-8601).
          example: 2026-05-14T09:30:00
        id:
          type: string
          format: uuid
          description: Comment id.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      required:
      - authorId
      - authorName
      - bugReportId
      - content
      - createdAt
      - id
...
    CreateCommentRequest:
      type: object
      description: "Request body of `POST /api/comments`."
      properties:
        content:
          type: string
          description: Comment text.
          example: This also happens in a private browser window.
          minLength: 1
        reportId:
          type: string
          format: uuid
          description: Id of the report to comment on.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      required:
      - content
      - reportId
    CreateCommentResponse:
      type: object
      description: "Response of `POST /api/comments`: the saved\
        \ comment."
      properties:
        authorId:
          type: string
          format: uuid
          description: Id of the author.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        authorName:
          type: string
          description: Display name of the author.
          example: Rachel Reporter
        content:
          type: string
          description: Comment text.
          example: This also happens in a private browser window.
        createdAt:
          type: string
          format: date-time
          description: When the comment was written (ISO-8601).
          example: 2026-05-14T09:30:00
        id:
          type: string
          format: uuid
          description: Id of the new comment.
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      required:
      - authorId
      - authorName
      - content
      - createdAt
      - id
...